% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_p.r
\name{stoufferMethod}
\alias{stoufferMethod}
\title{Merge p-values using stouffer methods with weights}
\usage{
stoufferMethod(p.values, weights)
}
\description{
Merge p-values using stouffer methods with weights
}
